+++ services/digitize/digitize_core.py
@@ -1,6 +1,6 @@
 from common.misc_utils import *
 from pathlib import Path
-from digitize.digitize_utils import get_utc_timestamp
+from common.misc_utils import get_utc_timestamp
 from digitize.models import JobStatus, DocStatus, OutputFormat
 from digitize.pdf_utils import get_pdf_page_count, get_document_page_count
 from digitize.doc_utils import convert_document_format
