+++ services/summarize/job_utils.py
@@ -6,8 +6,7 @@ Includes file validation, staging, directory initialization, and job operations.
 
 import os
 import shutil
-import uuid
-from datetime import datetime, timezone
+
 from pathlib import Path
 from typing import Optional, Tuple
 
@@ -116,110 +115,3 @@ def cleanup_staging_directory(job_id: str) -> None:
             logger.warning(f"Failed to clean up staging directory for job {job_id}: {e}")
 
 
-def create_job_record(
-    job_id: str,
-    document_name: str,
-    level: str = 'standard',
-    job_name: Optional[str] = None
-) -> SummarizeJob:
-    """
-    Create a new job record in the database.
-    
-    Args:
-        job_id: UUID for the job
-        document_name: Original filename
-        level: Summarization level (brief, standard, detailed)
-        job_name: Optional human-readable job name
-        
-    Returns:
-        Created SummarizeJob instance
-        
-    Raises:
-        Exception: If database insertion fails
-    """
-    job = SummarizeJob(
-        job_id=job_id,
-        job_name=job_name,
-        status='accepted',
-        submitted_at=datetime.now(timezone.utc),
-        document_name=document_name,
-        level=level,
-        job_type='direct',  # Will be updated to 'chunked' if needed during processing
-    )
-    
-    try:
-        with get_db_session() as session:
-            session.add(job)
-            session.commit()
-            session.refresh(job)
-        
-        logger.info(f"Created job record: {job_id}")
-        return job
-        
-    except Exception as e:
-        logger.error(f"Failed to create job record for {job_id}: {e}")
-        raise
-
-
-def get_job_by_id(job_id: str) -> Optional[SummarizeJob]:
-    """
-    Retrieve a job record by ID.
-    
-    Args:
-        job_id: UUID of the job
-        
-    Returns:
-        SummarizeJob instance or None if not found
-    """
-    try:
-        with get_db_session() as session:
-            job = session.query(SummarizeJob).filter(
-                SummarizeJob.job_id == job_id
-            ).first()
-            return job
-    except Exception as e:
-        logger.error(f"Failed to retrieve job {job_id}: {e}")
-        return None
-
-
-def update_job_status(
-    job_id: str,
-    status: str,
-    error: Optional[str] = None,
-    completed_at: Optional[datetime] = None
-) -> bool:
-    """
-    Update the status of a job.
-    
-    Args:
-        job_id: UUID of the job
-        status: New status (accepted, in_progress, completed, failed)
-        error: Error message if status is 'failed'
-        completed_at: Completion timestamp
-        
-    Returns:
-        True if update successful, False otherwise
-    """
-    try:
-        with get_db_session() as session:
-            job = session.query(SummarizeJob).filter(
-                SummarizeJob.job_id == job_id
-            ).first()
-            
-            if not job:
-                logger.warning(f"Job {job_id} not found for status update")
-                return False
-            
-            job.status = status
-            if error:
-                job.error = error
-            if completed_at:
-                job.completed_at = completed_at
-            
-            session.commit()
-            logger.info(f"Updated job {job_id} status to {status}")
-            return True
-            
-    except Exception as e:
-        logger.error(f"Failed to update job {job_id} status: {e}")
-        return False
